+++ src/Anneck/Game/Action/AbstractAction.php
@@ -5,33 +5,30 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  * ************************************************************************
- * Created at 04.03.15, 12:41 by 4nn3ck
+ * Created at 04.03.15, 16:28 by 4nn3ck
  * ************************************************************************
  */
 
 namespace Anneck\Game\Action;
 
-use Anneck\Game\GameInterface;
-use Anneck\Game\Item\ShopProduct;
-
 /**
- * The CreateShopProduct.
- *
- * @todo    Write PHPDoc for this class!
+ * The AbstractItemAction class serves all implementations as a base class.
  *
  * @since   0.0.1-dev
  *
  * @author  André Anneck <andreanneck73@gmail.com>
  */
-class CreateShopProduct extends AbstractAction
+abstract class AbstractAction implements ActionInterface
 {
     /**
-     * @param GameInterface $game
+     * One default toString implementation ...
+     *
+     * @return string the shortName of the class.
      */
-    public function applyOn(GameInterface $game)
+    public function __toString()
     {
-        $newItem = new ShopProduct($game);
-        $game->addItemToRegister($newItem);
-        $game->addScore(1);
+        $reClass = new \ReflectionClass($this);
+
+        return $reClass->getShortName();
     }
 }
